feat: wait for service worker before loading app:
before, the app started before the service worker had "claimed" -- only on first page load.

this meant that requests were not intercepted, or added to the cache -- unless you refreshed the page first.

also refactored the service worker to be in typescript, and set up proper precaching of all assets